refactor: split iron proxy resource backend
